+++ app/src/main/kotlin/com/urunkarpm/pingpin/receiver/AlarmReceiver.kt
@@ -51,9 +51,7 @@ class AlarmReceiver : BroadcastReceiver() {
             putExtra("portalUrl", portalUrl)
             addFlags(
                 Intent.FLAG_ACTIVITY_NEW_TASK or
-                        Intent.FLAG_ACTIVITY_CLEAR_TOP or
-                        Intent.FLAG_ACTIVITY_SINGLE_TOP or
-                        Intent.FLAG_ACTIVITY_REORDER_TO_FRONT
+                Intent.FLAG_ACTIVITY_NO_USER_ACTION
             )
         }